+++ crates/miden-testing/src/kernel_tests/tx/test_prologue.rs
@@ -13,7 +13,7 @@ use miden_lib::{
         memory::{
             ACCT_DB_ROOT_PTR, ACCT_ID_PTR, BLOCK_COMMITMENT_PTR, BLOCK_METADATA_PTR,
             BLOCK_NUMBER_IDX, CHAIN_COMMITMENT_PTR, INIT_ACCT_COMMITMENT_PTR, INIT_NONCE_PTR,
-            INPUT_NOTE_ARGS_OFFSET, INPUT_NOTE_ASSETS_HASH_OFFSET, INPUT_NOTE_ASSETS_OFFSET,
+            INPUT_NOTE_ARGS_OFFSET, INPUT_NOTE_ASSETS_COMMITMENT_OFFSET, INPUT_NOTE_ASSETS_OFFSET,
             INPUT_NOTE_ID_OFFSET, INPUT_NOTE_INPUTS_COMMITMENT_OFFSET, INPUT_NOTE_METADATA_OFFSET,
             INPUT_NOTE_NULLIFIER_SECTION_PTR, INPUT_NOTE_NUM_ASSETS_OFFSET,
             INPUT_NOTE_SCRIPT_ROOT_OFFSET, INPUT_NOTE_SECTION_PTR, INPUT_NOTE_SERIAL_NUM_OFFSET,
@@ -414,7 +414,7 @@ fn input_notes_memory_assertions(
         );
 
         assert_eq!(
-            read_note_element(process, note_idx, INPUT_NOTE_ASSETS_HASH_OFFSET),
+            read_note_element(process, note_idx, INPUT_NOTE_ASSETS_COMMITMENT_OFFSET),
             note.assets().commitment(),
             "note asset commitment should be stored at the correct offset"
         );